Add volume_attachment as a first-class object
- SDK v1 had Attachments as a first-class object which made our waiter
  very simple
- While this object was removed in v2, we made our own replacement
  inspired by the original v1 implementation

+++ src/bosh_aws_cpi/spec/integration/lifecycle_spec.rb
@@ -230,7 +230,7 @@ describe Bosh::AwsCloud::Cloud do
             expect(snapshot_tags['Name']).to eq 'deployment/cpi_spec/0/sdf'
 
           ensure
-            cpi.delete_snapshot(snapshot_id)
+            cpi.delete_snapshot(snapshot_id) if snapshot_id
             Bosh::Common.retryable(tries: 20, on: Bosh::Clouds::DiskNotAttached, sleep: lambda { |n, _| [2**(n-1), 30].min }) do
               cpi.detach_disk(instance_id, volume_id)
               true
